pyproject.toml
[tool.poetry]
name = "fastapi_sqlalchemy_async"
version = "0.1.0"
description = ""
authors = ["Marcelo Lino <mdslino@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
alembic = "^1.4.3"
SQLAlchemy = "1.4.0b1"
asyncpg = "^0.21.0"
psycopg2-binary = "^2.8.6"
pydantic = {extras = ["email"], version = "^1.7.3"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.2.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
